Refactor codec utils
