+++ bounded-collections/src/bounded_btree_map.rs
@@ -430,47 +430,10 @@ where
 macro_rules! codec_impl {
 	($codec:ident) => {
 		use super::*;
+		use crate::codec_utils::PrependCompactInput;
 		use $codec::{
 			Compact, Decode, DecodeLength, DecodeWithMemTracking, Encode, EncodeLike, Error, Input, MaxEncodedLen,
 		};
-
-		// Struct which allows prepending the compact after reading from an input.
-		pub(crate) struct PrependCompactInput<'a, I> {
-			pub encoded_len: &'a [u8],
-			pub read: usize,
-			pub inner: &'a mut I,
-		}
-
-		impl<'a, I: Input> Input for PrependCompactInput<'a, I> {
-			fn remaining_len(&mut self) -> Result<Option<usize>, Error> {
-				let remaining_compact = self.encoded_len.len().saturating_sub(self.read);
-				Ok(self.inner.remaining_len()?.map(|len| len.saturating_add(remaining_compact)))
-			}
-
-			fn read(&mut self, into: &mut [u8]) -> Result<(), Error> {
-				if into.is_empty() {
-					return Ok(());
-				}
-
-				let remaining_compact = self.encoded_len.len().saturating_sub(self.read);
-				if remaining_compact > 0 {
-					let to_read = into.len().min(remaining_compact);
-					into[..to_read].copy_from_slice(&self.encoded_len[self.read..][..to_read]);
-					self.read += to_read;
-
-					if to_read < into.len() {
-						// Buffer not full, keep reading the inner.
-						self.inner.read(&mut into[to_read..])
-					} else {
-						// Buffer was filled by the compact.
-						Ok(())
-					}
-				} else {
-					// Prepended compact has been read, just read from inner.
-					self.inner.read(into)
-				}
-			}
-		}
 
 		impl<K, V, S> Decode for BoundedBTreeMap<K, V, S>
 		where
@@ -549,9 +512,7 @@ mod test {
 	use crate::ConstU32;
 	use alloc::{vec, vec::Vec};
 	#[cfg(feature = "scale-codec")]
-	use scale_codec::{Compact, CompactLen, Decode, Encode, Input};
-	#[cfg(feature = "scale-codec")]
-	use scale_codec_impl::PrependCompactInput;
+	use scale_codec::{Compact, CompactLen, Decode, Encode};
 
 	fn map_from_keys<K>(keys: &[K]) -> BTreeMap<K, ()>
 	where
@@ -803,59 +764,6 @@ mod test {
 			[1, 2, 3, 4].into_iter().map(|k| (k, (k as u16) * 100)).try_collect().unwrap();
 
 		assert_eq!(Ok(b2), b1.try_map(|(_, v)| (v as u16).checked_mul(100_u16).ok_or("overflow")));
-	}
-
-	#[test]
-	#[cfg(feature = "scale-codec")]
-	fn prepend_compact_input_works() {
-		let encoded_len = Compact(3u32).encode();
-		let inner = [2, 3, 4];
-		let mut input = PrependCompactInput { encoded_len: encoded_len.as_ref(), read: 0, inner: &mut &inner[..] };
-		assert_eq!(input.remaining_len(), Ok(Some(4)));
-
-		// Passing an empty buffer should leave input unchanged.
-		let mut empty_buf = [];
-		assert_eq!(input.read(&mut empty_buf), Ok(()));
-		assert_eq!(input.remaining_len(), Ok(Some(4)));
-		assert_eq!(input.read, 0);
-
-		// Passing a correctly-sized buffer will read correctly.
-		let mut buf = [0; 4];
-		assert_eq!(input.read(&mut buf), Ok(()));
-		assert_eq!(buf[0], encoded_len[0]);
-		assert_eq!(buf[1..], inner[..]);
-		// And the bookkeeping agrees.
-		assert_eq!(input.remaining_len(), Ok(Some(0)));
-		assert_eq!(input.read, encoded_len.len());
-
-		// And we can't read more.
-		assert!(input.read(&mut buf).is_err());
-	}
-
-	#[test]
-	#[cfg(feature = "scale-codec")]
-	fn prepend_compact_input_incremental_read_works() {
-		let encoded_len = Compact(3u32).encode();
-		let inner = [2, 3, 4];
-		let mut input = PrependCompactInput { encoded_len: encoded_len.as_ref(), read: 0, inner: &mut &inner[..] };
-		assert_eq!(input.remaining_len(), Ok(Some(4)));
-
-		// Compact is first byte - ensure that it fills the buffer when it's more than one.
-		let mut buf = [0u8; 2];
-		assert_eq!(input.read(&mut buf), Ok(()));
-		assert_eq!(buf[0], encoded_len[0]);
-		assert_eq!(buf[1], inner[0]);
-		assert_eq!(input.remaining_len(), Ok(Some(2)));
-		assert_eq!(input.read, encoded_len.len());
-
-		// Check the last two bytes are read correctly.
-		assert_eq!(input.read(&mut buf), Ok(()));
-		assert_eq!(buf[..], inner[1..]);
-		assert_eq!(input.remaining_len(), Ok(Some(0)));
-		assert_eq!(input.read, encoded_len.len());
-
-		// And we can't read more.
-		assert!(input.read(&mut buf).is_err());
 	}
 
 	// Just a test that structs containing `BoundedBTreeMap` can derive `Hash`. (This was broken
