+++ bounded-collections/src/lib.rs
@@ -16,6 +16,7 @@ pub extern crate alloc;
 pub mod bounded_btree_map;
 pub mod bounded_btree_set;
 pub mod bounded_vec;
+pub(crate) mod codec_utils;
 pub mod const_int;
 pub mod weak_bounded_vec;
